requirements.txt
numpy
matplotlib
torch
torchvision
pyyaml
pathlib
opencv-python==3.4.2.17
opencv-contrib-python==3.4.2.17
